services/inference/tryon/requirements.txt
Pillow==10.4.0
httpx==0.27.0
boto3==1.34.144
aioredis==2.0.1
asyncpg==0.29.0
numpy==1.26.4
python-dotenv==1.0.1
sentry-sdk==2.10.0
# Face detection (lightweight, no GUI deps)
opencv-python-headless==4.10.0.84
# Uncomment to enable Replicate provider:
# replicate==0.30.2
# Uncomment to enable Segmind provider:
# (no extra deps needed)
